rekap kegiatan number_format

+++ database/migrations/2021_11_21_000900_create_rekap_kegiatan_table.php
@@ -26,12 +26,12 @@ class CreateRekapKegiatanTable extends Migration
             $table->integer('k_renja_pd');
             $table->integer('k_rka_pd');
             $table->integer('k_dpa_pd');
-            $table->integer('k_jumlah')->nullable();
-            $table->integer('pa_rkpd');
-            $table->integer('pa_ranwal_renja');
-            $table->integer('pa_renja_pd');
-            $table->integer('pa_rka_pd');
-            $table->integer('pa_dpa_pd');
+            $table->bigInteger('k_jumlah')->nullable();
+            $table->bigInteger('pa_rkpd');
+            $table->bigInteger('pa_ranwal_renja');
+            $table->bigInteger('pa_renja_pd');
+            $table->bigInteger('pa_rka_pd');
+            $table->bigInteger('pa_dpa_pd');
             $table->string('ket')->nullable();
             $table->timestamps();
             $table->softDeletes();
